create structure for a ram cache manager
