+++ cvmfs/CMakeLists.txt
@@ -34,6 +34,8 @@ set (CVMFS_CLIENT_SOURCES
   backoff.cc backoff.h
   bigvector.h
   cache.cc cache.h
+  cache_posix.cc cache_posix.h
+  cache_ram.cc cache_ram.h
   catalog.cc catalog.h
   catalog_counters.cc catalog_counters.h catalog_counters_impl.h
   catalog_mgr.h catalog_mgr_impl.h
