+++ cvmfs/cache_ram.cc
@@ -1,0 +1,11 @@
+/**
+ * This file is part of the CernVM File System.
+ */
+#include "cvmfs_config.h"
+#include "cache_ram.h"
+
+using namespace std;  // NOLINT
+
+namespace cache {
+
+}  // namespace cache
