+++ cvmfs/cvmfs.cc
@@ -72,7 +72,7 @@
 #include "authz/authz_session.h"
 #include "auto_umount.h"
 #include "backoff.h"
-#include "cache.h"
+#include "cache_posix.h"
 #include "catalog_mgr_client.h"
 #include "clientctx.h"
 #include "compat.h"
